Module streaming
Added module streaming, still unfinished, it outputs a file called 'ModuleRaw.txt' where you can view the streamed data, this is obviously temporary.

The binary is streamed using 3KB chunks per stream.

This is unfinished as we still need a nicer way to load the binary onto the server's memory.